requirements.txt
python-frontmatter
jinja2
pyromark